feat(sagemaker): Support remote access deep link for Kiro (#2263)
## Problem
When a user initiates remote access to their SageMaker Space from a web
page (such as from the SageMaker AI web portal), known as the "deep
link" feature, depending on the user's browser, they can choose to open
the link in Kiro instead of in VS Code. When the user opens the link in
Kiro, the remote connection logic checks for the Microsoft Remote - SSH
extension which can't be installed in Kiro, resulting in the connection
attempt being aborted.

## Solution
In the deep link case, apply the same branching logic that is used when
initiating remote access from the AWS Toolkit, in order to use the
SageMaker Kiro SSH extension instead.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

Co-authored-by: Laxman Reddy <141967714+laileni-aws@users.noreply.github.com>