fix(sagemaker): Fix path resolution for `kiro` executable (#2271)
## Problem
When a user tries to connect to a SageMaker Space in Kiro on Mac, the
remote workspace window is not able to be opened unless the user has
`kiro` on their `$PATH`. Ideally, we want to directly find the `kiro`
executable rather than rely on the user's `$PATH`. The problem is that
the existing path finding logic is not able to locate the `kiro`
executable on Mac because it's actually named `code` (though confirmed
to be named `kiro` on Windows and Linux).

## Solution
Update the path finding logic to fall back to `code` executables in
relevant Kiro application directories if the `kiro` executable is not
found. Invoking `kiro` on the `$PATH` will only be done as a last
resort.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.