fix(sagemaker): Fix host identification issue after restarting a Space (#8249)
## Problem
Users will face a host identification issue when reconnecting to a Space
after restarting it, if the Space name has capital letters. This occurs
because the Space's hostname is written to the known_hosts file in
lowercase, but hostname matching during reconnection is case-sensitive.

## Solution
When removing the Space hostname from the known_hosts file, ensure to
remove lines that match the lowercase variant of the Space hostname.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

Co-authored-by: Laxman Reddy <141967714+laileni-aws@users.noreply.github.com>