fix(q): Remove Q Show Logs options from non Q views
**Description**

Noticed that if Q extension is installed alongside AWS Toolkit,
the Show Logs option redirecting to Q logs were showing up in
the toolkit views as well. This does not make sense, especially
for SMUS as it is misleading.

Change now shows this only for amazonq views similar to the learn more
option.

**Motivation**

Remove misleading show logs option from non Q views

**Testing Done**

Tested locally that it no longer shows up in Toolkit but
shows in Q.