feat(smus): Update region selector to allow matching on region code (#8324)
## Problem
When users sign in to SageMaker Unified Studio with IAM credentials, if
they use the step-by-step wizard to add a new profile, when the user is
prompted to select a region, typing in a region code such as `us-east-1`
to filter the results does not show expected results because the
QuickPick matches only on the item labels, not on their description.

<img width="602" height="195" alt="image"
src="https://github.com/user-attachments/assets/695f064f-b6bd-4de6-bf9a-71741b3b566d"
/>

## Solution
Update the region selection QuickPick to also match options according to
the description (in addition to the label).

<img width="598" height="326" alt="image"
src="https://github.com/user-attachments/assets/902ad2ed-19f3-4be0-bd6e-3b768745f9ba"
/>

---

Tests are not necessary for this change, given that we are just changing
a config option of the quick pick menu.

Not adding a changelog because this updates unreleased functionality in
the `feature/smus` branch.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

---------

Co-authored-by: Laxman Reddy <141967714+laileni-aws@users.noreply.github.com>