fix(cloudformation): handle telemetry setting in upgrade path case where setting is not registered (#8343)
## Problem
When the extension is upgraded and user does not do a full restart, it
is possible for new cloudformation telemetry setting to not be
registered yet, resulting in error trying to save it.

## Solution
Skip the save and do it the next time the extension starts.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

---------

Co-authored-by: Akila Tennakoon <tennakoo@amazon.com>
Co-authored-by: manodnyab <66754471+manodnyab@users.noreply.github.com>