feat(stepfunctions): Allow redriving a failed execution in ExecutionDetails Tab (#8186)
## Problem

Currently in the StepFunctions ExecutionDetails Tab, we show the
`redrive` button for a Standard failed execution, but clicking the
button will cause the webview to crash.

## Solution

Implemented a message handler for `redrive` messages passed from the JS
bundle, which then opens a new ExecutionDetails Webview with the
ExecutionArn.


![RedriveButton](https://github.com/user-attachments/assets/5c6cb2e6-4ae2-46f8-af5d-1eeaa7eb1d0e)


---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

---------

Co-authored-by: Zelin Zhou <zelzhou@amazon.com>
Co-authored-by: Laxman Reddy <141967714+laileni-aws@users.noreply.github.com>