fix(lambda): remove version publish for every lambda deploy (#8400)
## Problem
Fixes #8372.

Currently, a deploy done using the lambda client will create a new
published version of the function without any customer input. This is an
issue because a) it is not the expected action for a deploy and b)
version creation can take a long time and be expensive for Lambda
Managed Instances customers.

The cases in which this happened were: 
1. Quick deploy through either the popup or the toolkit explorer
2. Direct upload where the user clicks on the node and selects the
upload action
3. Remote invoke hot reloading, which will auto deploy code changes.

Open to reasons to keep this, but I think that versioning isn't
necessary in those cases and can end up being confusing. The original PR
that added this was in 2020 (#1269), and based on that PR it seems like
the original implementer was testing things out.

## Solution
Remove the `Publish: true` field from the `UpdateFunctionCodeCommand`.
Now, all deploys are simply updating the code.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

Co-authored-by: invictus <149003065+ashishrp-aws@users.noreply.github.com>